Removed the old CRA client
