--- client/README.md
@@ -1,3 +1,0 @@
-# Chat application client
-
-Made using React/Typescript
