+++ client/package.json
@@ -1,64 +1,39 @@
 {
-  "name": "chatzilla-client",
-  "version": "1.2.0",
+  "name": "client2",
   "private": true,
-  "author": "Poujhit <poujhit.dev@gmail.com>",
-  "dependencies": {
-    "@material-ui/core": "^4.11.3",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router": "^5.1.13",
-    "@types/react-router-dom": "^5.1.7",
-    "axios": "^0.21.1",
-    "emoji-mart": "^3.0.1",
-    "formik": "^2.2.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-linkify": "^1.0.0-alpha",
-    "react-query": "^3.13.4",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-scrollable-feed": "^1.3.0",
-    "react-spinners": "^0.10.6",
-    "socket.io-client": "^4.0.1",
-    "typescript": "^4.1.2",
-    "web-vitals": "^1.0.1",
-    "zustand": "^3.3.3"
-  },
+  "version": "2.0.0",
+  "type": "module",
   "scripts": {
-    "start": "react-scripts start",
-    "start:for:electron": "cross-env BROWSER=none react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite --port 3000",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "@emoji-mart/data": "^1.1.2",
+    "@emoji-mart/react": "^1.1.1",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/material": "^5.11.6",
+    "axios": "^1.2.5",
+    "emoji-mart": "^5.5.2",
+    "formik": "^2.2.9",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-linkify": "^1.0.0-alpha",
+    "react-query": "^3.39.3",
+    "react-router-dom": "^6.8.0",
+    "react-spinners": "^0.13.8",
+    "socket.io-client": "^4.5.4",
+    "tss-react": "^4.5.2",
+    "zustand": "^4.3.2"
   },
   "devDependencies": {
-    "@types/emoji-mart": "^3.0.8",
+    "@types/node": "^18.11.18",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
     "@types/react-linkify": "^1.0.1",
-    "cross-env": "^7.0.3"
+    "@vitejs/plugin-react": "^3.0.0",
+    "typescript": "^4.9.3",
+    "vite": "^4.0.0"
   }
 }
