--- client/public/index.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/icon.png" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
-    <meta
-      name="description"
-      content="A chatroom application that lets you chat with others on the fly."
-    />
-    <meta property="og:type" content="website" />
-    <meta property="og:title" content="Chatzilla" />
-    <meta
-      property="og:description"
-      content="A chatroom application that lets you chat with others on the fly."
-    />
-    <meta property="og:image" content="%PUBLIC_URL%/chatzilla.jpeg" />
-    <meta property="og:image:secure" content="%PUBLIC_URL%/chatzilla.jpeg" />
-
-    <meta property="og:url" content="https://chatzilla-web.netlify.app" />
-
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
-
-    <title>Chatzilla</title>
-  </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-
-    <div id="root"></div>
-  </body>
-</html>
