--- client/src/containers/Chat/MessagePortion.tsx
@@ -1,30 +1,0 @@
-import React from 'react';
-import ScrollableFeed from 'react-scrollable-feed';
-
-import { Message } from './ChatScreen';
-import useChatScreenStyles from './ChatScreenStyles';
-import EachMessage from 'components/messages/EachMessage';
-
-interface MessagePortionProps {
-  messages: Message[];
-  name: string;
-}
-
-const MessagePortion: React.FC<MessagePortionProps> = (props) => {
-  const classes = useChatScreenStyles();
-  return (
-    <ScrollableFeed forceScroll={true} className={classes.MessagePortion}>
-      {props.messages.map((eachmessage, index) => {
-        return (
-          <EachMessage
-            key={index}
-            message={eachmessage}
-            username={props.name}
-          />
-        );
-      })}
-    </ScrollableFeed>
-  );
-};
-
-export default MessagePortion;
