--- client/src/index.tsx
@@ -1,19 +1,0 @@
-import ReactDOM from 'react-dom';
-import './index.css';
-import { QueryClientProvider, QueryClient } from 'react-query';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
-
-const clientQuery = new QueryClient();
-
-ReactDOM.render(
-  <QueryClientProvider client={clientQuery}>
-    <App />
-  </QueryClientProvider>,
-  document.getElementById('root')
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
