+++ client/src/views/chatView/MessageInputPortion.tsx
