+++ client/src/views/chatView/MessagePortion.tsx
