--- client2/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "client2",
-  "private": true,
-  "version": "2.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite --port 3000",
-    "build": "tsc && vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@emoji-mart/data": "^1.1.2",
-    "@emoji-mart/react": "^1.1.1",
-    "@emotion/react": "^11.10.5",
-    "@emotion/styled": "^11.10.5",
-    "@mui/material": "^5.11.6",
-    "axios": "^1.2.5",
-    "emoji-mart": "^5.5.2",
-    "formik": "^2.2.9",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-linkify": "^1.0.0-alpha",
-    "react-query": "^3.39.3",
-    "react-router-dom": "^6.8.0",
-    "react-spinners": "^0.13.8",
-    "socket.io-client": "^4.5.4",
-    "tss-react": "^4.5.2",
-    "zustand": "^4.3.2"
-  },
-  "devDependencies": {
-    "@types/node": "^18.11.18",
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^18.0.9",
-    "@types/react-linkify": "^1.0.1",
-    "@vitejs/plugin-react": "^3.0.0",
-    "typescript": "^4.9.3",
-    "vite": "^4.0.0"
-  }
-}
