--- client2/src/App.css
@@ -1,6 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Rubik&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Rubik:wght@700&display=swap');
-
-.App {
-  text-align: center;
-}
