--- client2/src/components/messages/EachMessage.tsx
@@ -1,97 +1,0 @@
-import React from 'react';
-
-import eachMessageStyles from './EachMessageStyles';
-import { Message } from 'views/chatView/ChatScreen';
-import Linkify from 'react-linkify';
-import { PropsWithChildren } from 'react';
-
-interface EachMessageProps {
-  message: Message;
-  username: string;
-}
-
-const LinkifyWrapper: React.FC<PropsWithChildren> = ({ children }) => (
-  <Linkify
-    componentDecorator={(decoratedHref, decoratedText, key) => (
-      <a
-        target='_blank'
-        style={{ color: 'white' }}
-        href={decoratedHref}
-        key={key}
-        rel='noreferrer'
-      >
-        {decoratedText}
-      </a>
-    )}
-  >
-    {children}
-  </Linkify>
-);
-
-const EachMessage: React.FC<EachMessageProps> = (props) => {
-  const classes = eachMessageStyles();
-  let isSentByCurrentUser = false;
-  const trimmedName = props.username.trim().toLowerCase();
-
-  if (props.message.user === trimmedName) {
-    isSentByCurrentUser = true;
-  }
-
-  if (isSentByCurrentUser) {
-    return (
-      <div
-        className={classes.myMessageContainer}
-        style={{
-          justifyContent: 'flex-end',
-          marginRight: '10px',
-        }}
-      >
-        <p className={classes.userNameText}>{props.message.user}</p>
-        <div className={classes.myMessageTextBox}>
-          <p className={classes.messageText}>
-            <LinkifyWrapper>{props.message.text}</LinkifyWrapper>
-          </p>
-        </div>
-      </div>
-    );
-  } else {
-    if (props.message.user === 'Admin') {
-      return (
-        <div
-          className={classes.myMessageContainer}
-          style={{
-            justifyContent: 'center',
-          }}
-        >
-          <div
-            className={classes.otherMessageTextBox}
-            style={{
-              backgroundColor: '#2263A5',
-            }}
-          >
-            <p className={classes.messageText}>
-              <LinkifyWrapper>{props.message.text}</LinkifyWrapper>
-            </p>
-          </div>
-        </div>
-      );
-    } else {
-      return (
-        <div
-          className={classes.myMessageContainer}
-          style={{
-            justifyContent: 'flex-start',
-            marginLeft: '10px',
-          }}
-        >
-          <div className={classes.otherMessageTextBox}>
-            <p className={classes.messageText}>{props.message.text}</p>
-          </div>
-          <p className={classes.userNameText}>{props.message.user}</p>
-        </div>
-      );
-    }
-  }
-};
-
-export default EachMessage;
