--- client2/src/print.ts
@@ -1,5 +1,0 @@
-const print = (data: any): void => {
-  console.log(data);
-};
-
-export default print;
