--- client2/src/theme/Theme.ts
@@ -1,17 +1,0 @@
-import { createTheme } from '@material-ui/core';
-
-export const theme = createTheme({
-  typography: {
-    fontFamily: 'Rubik, sans-serif',
-  },
-
-  palette: {
-    type: 'dark',
-    primary: {
-      main: '#6586e6',
-    },
-    error: {
-      main: '#f65164',
-    },
-  },
-});
